azure-function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
PyMuPDF==1.24.7
azure-storage-blob==12.20.0
azure-identity==1.16.1
